+++ RELEASE_NOTES.md
@@ -1,3 +1,10 @@
+### Version 0.2 (March 18, 2011)
+
+* Added new task `gaeExplodeWar`.
+* Cut dependency to exploded WAR directory for `gaeVersion` task.
+* Task `gaeRun` will use exploded WAR directory by default. This behavior can be overriden by the convention property
+`warDir`.
+
 ### Version 0.1 (March 17, 2011)
 
 * Initial release.